encapp: script: move ffprobe code to separate file
We will use the file for more ffmpeg-related code.
